remove URLAddressableTable layer
