+++ ui/components/ReconciledObjectsTable.tsx
@@ -10,7 +10,7 @@ import {
 import { formatURL, objectTypeToRoute } from "../lib/nav";
 import { NoNamespace } from "../lib/types";
 import { addKind, makeImageString, statusSortHelper } from "../lib/utils";
-import {
+import FilterableTable, {
   filterByStatusCallback,
   filterByTypeCallback,
   filterConfig,
@@ -19,7 +19,6 @@ import ImageLink from "./ImageLink";
 import KubeStatusIndicator, { computeMessage } from "./KubeStatusIndicator";
 import Link from "./Link";
 import RequestStateHandler from "./RequestStateHandler";
-import URLAddressableTable from "./URLAddressableTable";
 
 export interface ReconciledVisualizationProps {
   className?: string;
@@ -73,7 +72,7 @@ function ReconciledObjectsTable({
 
   return (
     <RequestStateHandler loading={isLoading} error={error}>
-      <URLAddressableTable
+      <FilterableTable
         filters={initialFilterState}
         className={className}
         fields={[
